move: move old files to old folder
